# [Job Log perf] Use rAF to call throttleSetNodes